requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy<2
